Fix GMR break with new psutil versions
psutil 2.x and above has a lot of API changes as described in:
https://github.com/giampaolo/psutil/blob/master/HISTORY.rst

So we should work correctly with both old and new psutil versions
by using a version check and use the correct method/attributes.

Note that we cannot remove the g-r cap until all the projects have
switched to oslo.reports (OR we fix oslo-incubator GMR code and have
sync'ed all the consuming projects). So this review is the first
step in a long process of switching all consumers of GMR to use
what we have in oslo.reports.


Depends-On: I0f07858e96ea3baf46f8a453e253b9ed29c7f7e2
Depends-On: I33bd2d9dff9cb7dc1a50177db7286b7317966784

Closes-Bug: #1430231
Change-Id: I52bc6898d098f599c12204fc6d6eb269e2c884cb