Setup translations
To start translation, we need to initially import the
translation file - and place it at the proper place so that
the usual CI scripts can handle it.

The proper place is for all python projects
$PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports will be done by the OpenStack Proposal bot.

This patch is needed by https://review.openstack.org/210774 .

Note that this project currently has not a single translatable string.
It needs to use oslo.i18n to mark strings as translatable.

Change-Id: I8672716071d3874411c7a0021ee71816691b0fcb