Document workaround for AccessDenied error
On SELinux-enabled platforms it is possible for the report process
to fail with an AccessDenied error when it tries to read information
about the process being debugged.  Per [1], the recommended solution
is to temporarily disable SELinux during debugging and then turn it
on again once the report has completed successfully.

1: https://bugzilla.redhat.com/show_bug.cgi?id=1292787

Change-Id: Ic12d5658858bb085448e1b437b548111d3c79583
Closes-Bug: 1756044