Update unit test to reflect latest oslo.config
Because we added the default config_source opt in oslo.config, it
is now showing up in the GMR output and needs to be included in the
expected content.  However, for lower constraints testing we also
need to continue to support the older versions without config_source,
so the target_str is built dynamically based on whether the opt
exists.

Change-Id: I20b01285000ce84b69e8bd3cee8078edf03616f8