Support integer keys of dicts in text serialization
When the config contains a DictOpt that accepts integer keys, the Guru
Meditation Report fails to serialize. Even if current OpenStack code
might not contain such dicts, it's possible that downstream add such
options. Therefore, we change the code to support both dicts with only
integer keys and - just in case - mixed-key dicts.

Change-Id: I44343a8c306c96fc8dc078a76e744cf8b897d8d8