[stable-only] Cap setuptools <71.0.0
py38 jobs (on ubuntu-focal) started to fail (see the relevant github
issue [1]) due to recent virtualenv release (20.26.4; which bundles
setuptools).

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

Furthermore, lower-constraints jobs are not supported anymore and are
removed from most of the branches, so let's remove here as well to
unblock the gate.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: Ib2d3ff7dbcbbaf8c5bc392f77091bea0a6c25002